[101] Starter code.
[101] Making AppBar something you add in 102.

[Model] Adding model from the start.

[103] Adding fonts.

[103] Supplemental files.

[101] Correcting supplemental files.

[101] Copy correction.

[101] Minor renaming.

[101] Update for Dart 2.

[101] Missing dependency.

[101] Dart 2.

[101] Dart 2.

[101] Dart 2.

[101] Dart 2.

[101] Correcting model.

[Meta] README file.

[Meta] Clarification.

[101] Copy correction.

[104] Updating data.

[101] Correcting widget class. (#23)

Add link to actual codelab in the README

[Meta] README correction.

[101] README specificity.

Add the slanted_menu.png to assets for 101 branch so its ready when needed. (#44)

[101] TODOs for starter code (#50)

[101] Added future TODOs to 101 starter

[101] Included lines to ignore .iml files and files in idea/ (#67)

[101] Remove 104 comments

[All] Icons for Android and iOS. (#69)

[101] Add .vscode/ to gitignore, stop tracking ignored files (#71)

Removes files that should not be tracked because they've been ignored by gitignore rules (but have yet to be removed from the repository itself). Also adds the .vscode folder to gitignore, for users who might use VSCode as their development environment.